+++ src/Luxor.jl
@@ -154,7 +154,7 @@ export Drawing,
     textoutlines, textcurve, textcentred, textcentered,
     textright, textcurvecentred, textcurvecentered,
     get_fontsize, textwrap, textlines, splittext, textbox,
-    texttrack, textplace,
+    texttrack, textplace, textfit,
 
     setcolor, setopacity, sethue, setgrey, setgray,
     randomhue, randomcolor, @setcolor_str,